requirements.txt
# Core dependencies for Telegram bot
python-telegram-bot==20.7
requests==2.31.0

# Database
sqlite3

# Web framework
Flask==3.0.0
Flask-CORS==4.0.0

# Date and time handling
python-dateutil==2.8.2

# Environment variables
python-dotenv==1.0.0

# Optional: Google Sheets integration
gspread==5.12.0
oauth2client==4.1.3

# Development dependencies
pytest==7.4.3
black==23.11.0
flake8==6.1.0

# Deployment
gunicorn==21.2.0